requirements.txt
Django
psycopg2-binary
django-extensions
django-filter
django-widget-tweaks
django-cors-headers
djangorestframework
djangorestframework-simplejwt
drf-yasg
celery
redis
django-celery-beat
django-celery-results
Pillow
beautifulsoup4
requests
cryptography
python-dotenv
django-environ
django-htmx
django-allauth
django-compressor
whitenoise
gunicorn
fido2
phonenumbers
drf-spectacular
django-ratelimit
django-validators
easy-thumbnails
python-magic
